Split get_rq_admin_context into separate function so admin can use rq views inside
For django 2.2: https://docs.djangoproject.com/en/2.2/ref/contrib/admin/#overriding-admin-templates

Overriding index.html (easier to do in 3.0 since app list is in separate template):

{% extends "admin/index.html" %}
{% load i18n static %}

{% comment %}
This override gets better in django 3.0 where it's app_list.html %}

This template only loads if django_rq_dashboard is enabled

if 'django_rq_dashboard' in settings.INSTALLED_APPS:
    admin.site.index_template = 'admin/pg_index.html'
{% endcomment %}

{% block content %}
<div id="content-main">
{% if app_list %}
    {% for app in app_list %}
        <div class="app-{{ app.app_label }} module">
        <table>
        <caption>
            <a href="{{ app.app_url }}" class="section" title="{% blocktrans with name=app.name %}Models in the {{ name }} application{% endblocktrans %}">{{ app.name }}</a>
        </caption>
        {% for model in app.models %}
            <tr class="model-{{ model.object_name|lower }}">
            {% if model.admin_url %}
                <th scope="row"><a href="{{ model.admin_url }}">{{ model.name }}</a></th>
            {% else %}
                <th scope="row">{{ model.name }}</th>
            {% endif %}

            {% if model.add_url %}
                <td><a href="{{ model.add_url }}" class="addlink">{% trans 'Add' %}</a></td>
            {% else %}
                <td>&nbsp;</td>
            {% endif %}

            {% if model.admin_url %}
                {% if model.view_only %}
                <td><a href="{{ model.admin_url }}" class="viewlink">{% trans 'View' %}</a></td>
                {% else %}
                <td><a href="{{ model.admin_url }}" class="changelink">{% trans 'Change' %}</a></td>
                {% endif %}
            {% else %}
                <td>&nbsp;</td>
            {% endif %}
            </tr>
        {% endfor %}
        </table>
        </div>
    {% endfor %}
{% else %}
    <p>{% trans "You don't have permission to view or edit anything." %}</p>
{% endif %}
{% include "rq/modules.html" %}
</div>
{% endblock %}

Admin site: https://docs.djangoproject.com/en/2.2/ref/contrib/admin/#overriding-default-admin-site

from django.contrib.admin import AdminSite

class PeergradeAdminSite(AdminSite):
    login_template = 'admin/pg_admin_login.html'  # your template override
    if 'django_rq_dashboard' in settings.INSTALLED_APPS:
        index_template = 'admin/pg_index.html'

    @never_cache
    def index(self, request, extra_context=None):
        """Override to pass extra context to django-rq-dashboard

        https://github.com/django/django/blob/2.2.14/django/contrib/admin/sites.py#L486
        """
        from django_rq_dashboard.utils import get_rq_admin_context

        app_list = self.get_app_list(request)

        context = {
            **self.each_context(request),
            'title': self.index_title,
            'app_list': app_list,
            **(extra_context or {}),
            **get_rq_admin_context(),
        }

        request.current_app = self.name

        return TemplateResponse(
            request, self.index_template or 'admin/index.html', context
        )
